requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.21
google-search-results>=2.4.2
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
schedule>=1.2.0
email-validator>=2.1.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
aiofiles==23.2.1
jinja2==3.1.3

